+++ lib/async_cubits.dart
@@ -6,7 +6,5 @@ export 'src/async_cubit.dart';
 export 'src/async_cubits_logger.dart';
 export 'src/async_value.dart';
 export 'src/future_cubit.dart';
-export 'src/future_with_args_cubit.dart';
 export 'src/mutation_cubit.dart';
-export 'src/refresh_stream_future_cubit_mixin.dart';
 export 'src/stream_cubit.dart';