+++ test/src/mocks.dart
@@ -9,31 +9,37 @@ class TestFutureCubit extends FutureCubit<String> {
 
   int futureCalls = 0;
 
-  @override
-  Future<String> future() async {
+  Future<String> _future() async {
     futureCalls++;
     if (throwsError) {
       throw Exception('error');
     }
     return 'data';
   }
+
+  Future<void> load() => performLoad(_future);
+
+  Future<void> refresh() => performRefresh(_future);
 }
 
-class TestFutureWithArgsCubit extends FutureWithArgsCubit<String, String> {
+class TestFutureWithArgsCubit extends FutureCubit<String> {
   TestFutureWithArgsCubit({this.throwsError = false});
 
   final bool throwsError;
 
   int futureCalls = 0;
 
-  @override
-  Future<String> future(String args) async {
+  Future<String> _future(String args) async {
     futureCalls++;
     if (throwsError) {
       throw Exception('error');
     }
     return 'data';
   }
+
+  Future<void> load(String args) => performLoad(() => _future(args));
+
+  Future<void> refresh(String args) => performRefresh(() => _future(args));
 }
 
 class TestStreamCubit extends StreamCubit<int> {